refactor: 쿼리 리포지토리용 이넘 이동
