+++ src/main/java/com/moyeoit/context/club/infra/query/ClubActivityType.java
@@ -1,27 +1,27 @@
-package com.moyeoit.context.club.domain.entity.enums;
+package com.moyeoit.context.club.infra.query;
 
 import java.util.stream.Stream;
 import lombok.Getter;
 
 
 @Getter
-public enum Way {
+public enum ClubActivityType {
     ONLINE("온라인"),
     OFFLINE("오프라인");
 
     private final String description;
 
-    Way(String description) {
+    ClubActivityType(String description) {
         this.description = description;
     }
 
-    public static Way fromString(String text) {
+    public static ClubActivityType fromString(String text) {
         if (text == null) {
             return null;
         }
 
-        return Stream.of(Way.values())
-                .filter(way->way.description.equalsIgnoreCase(text))
+        return Stream.of(ClubActivityType.values())
+                .filter(clubActivityType -> clubActivityType.description.equalsIgnoreCase(text))
                 .findFirst()
                 .orElseThrow(() -> new IllegalArgumentException("Invalid way: " + text));
     }
